meta: add tests/Ciphers; change ciphers dir to Ciphers for consistency
