+++ Ciphers/XORCipher.php
