+++ Ciphers/monoAlphabeticCipher.php
