+++ tests/Ciphers/MorseCodeTest.php
@@ -3,12 +3,12 @@
 use function PHPUnit\Framework\assertEquals;
 use PHPUnit\Framework\TestCase;
 
-require_once __DIR__ . '/../vendor/autoload.php';
-require_once __DIR__ . '/../ciphers/morseCode.php';
+require_once __DIR__ . '/../../vendor/autoload.php';
+require_once __DIR__ . '/../../Ciphers/MorseCode.php';
 
 class MorseCodeTest extends TestCase
 {
-    public function testCaesarCipher()
+    public function testMorseCodeCipher()
     {
         assertEquals('TEST 123', decode(encode('TEST 123')));
     }
